requirement.txt
Pillow==3.1.1
pytesseract==0.1.6
